+++ secrover/constants.py
@@ -1,4 +1,1 @@
-from pathlib import Path
-
-VERSION = "0.2"
-REPOS_FOLDER = Path("repos")
+VERSION = "0.3"