+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3] - 2025-10-11
+- core: pull git repos if already cloned
+- core: permit to change repos folder path
+- export: add suport for remote export of reports with rclone
+- docker: add cron support with supercronic
+- docker: fix opengrep not found by Python subprocess
+
 ## [0.2] - 2025-10-10
 - core: only clone selected git branch
 - core: update deps